refactor(experience): decompose disambiguateCompanyTitle into orchestrator + phase helpers (#412)
Move the 398-line `disambiguateCompanyTitle` (fallow CRAP ~82 CRITICAL) out of
`experience.ts` into a new sibling module `experience-disambiguate.ts`, split
into a thin orchestrator over named, single-purpose phase helpers:

  stripLeadingSectionHeaders  (#310 leading section-header strip)
  splitHeaderSegments         (delimiter / role-comma split, `via` tag)
  mapSegmentsToFields         → mapWithCompanyMatch / mapWithoutCompanyMatch
                                → mapTitleFirst (#372)
  recoverLocation             → rescueTeamLocation (#325 team→location)
  cleanFieldArtifacts         (steps 4/4b/5)
  applyTrailingCompanyGuard   (#310 step 6)

Pure refactor — no parser behavior changes. Control flow and all guard
conditions are preserved verbatim; every phase's "why" doc comment moves with
its code. Only `disambiguateCompanyTitle` is exported (imported by
`experience.ts`); all predicates/regexes stay module-private.

Result: orchestrator no longer flagged by fallow (CRAP < 30, drops off the
CRITICAL list). Remaining per-helper flags are cyclomatic-only, non-critical,
and permitted by the issue: mapWithoutCompanyMatch (22), recoverLocation (16),
stripLocationSuffix (15, unchanged/moved verbatim).

Verify gate green: typecheck, eslint, full suite (1698 passed), build. Zero
`*.expected.json` golden-snapshot edits and corpus-roundtrip green confirm
behavior identity. experience.ts drops 837 → 100 lines.

Closes #412
